ClusterCommands interface is implemented in BinaryJedis.
